refactor(studio): extract table capping logic into useCappedTables hook
